Moved daily state to daily lib
