+++ client/src/routes/Session/IntroPortal.tsx
@@ -35,7 +35,8 @@ import {SessionStackProps} from '../../lib/navigation/constants/routes';
 import {SPACINGS} from '../../common/constants/spacings';
 import Counter from './components/Counter/Counter';
 import useSessionExercise from './hooks/useSessionExercise';
-import {participantsAtom, sessionAtom} from './state/state';
+import {sessionAtom} from './state/state';
+import {participantsAtom} from '../../lib/daily/state/state';
 import useLeaveSession from './hooks/useLeaveSession';
 import VideoBase from './components/VideoBase/VideoBase';
 import useIsSessionHost from './hooks/useIsSessionHost';
